requirements.txt
requests==2.0.0
jsonpatch==0.4
pytest==2.3.5
docopt==0.6.1
PyYAML==3.10
six=1.4.1
